Separate connector tests to a another file
